Improve receiver visibility on initial spawn
This works around the client rendering bug for parented IO entities by spawning the RF receiver at the desired world position before parenting it. This allows clients in the vicinity to render it correctly (until the vehicle moves away from there), but does not address the issue when it's networked to new clients who were initially outside the network range, or to clients who reconnect.